Add main function to hooks
